staging: zram: Rate limit memory allocation errors
If an error occurs allocating memory for zram we will not be able
to fullfill the request to store the page in zram. The swap subsystem
still continues to try to swap out pages to zram even when this error
occurs since there is currently no facility to stop the swap subsystem
from swapping out during such errors. This can cause the system to be
overflowed with logging errors.

Reduce the amount of logging to prevent the kernel log
from being filled with these error messages.

Change-Id: I54b920337749ece59d9ca78fa8b29345ec7b976b
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
(cherry picked from commit c43dbe4267c15f848c2994abefb13140a8825508 for re-application onto zram-linux-v3.18.57. Rewrite the patch to not have race conditions with zram_rs_time.)