Disable High Memory to Slightly Speed Up the Kernel
The kernel has to manipulate the page tables to make high memory accessible from kernel space. To avoid this and speed up the kernel, we tweak the memory areas.

We've had to change CONFIG_VMALLOC_RESERVE from 400MB to 128MB. This seems OK because the drivers in the kernel don't seem to use more than 100MB of vmalloc.

We've also had to subtract 224MB from PAGE_OFFSET to make things fit. Userspace shouldn't mind too much that it's missing 224MB (~7%) of virtual address space.