Disable the Custom Pantech Panic Handler & Logging
The custom Pantech panic handler and logging code is very half baked.

It tries to display a blue screen, but most of the time it doesn't finish drawing. The blue screen indicates that the power key can be used to reset the system, but it doesn't work and the system stays on until the battery is removed. We also don't have the tools to retrieve the logged errors. The addition of the ram-console makes the Pantech error logging redundant too.

So, disable the Pantech panic handler/logger. Maybe if we have a panic, the regular mechanisms like the kernel's built-in reboot-on-panic and the watchdog can actually have a chance to reboot the system.