tki-melfas: Allow the touch-key interface's search button to function as a power button.
The search key can be used to wake up the phone whether the system is suspended or not. The touch-key interface is kept powered-up by default at all times to support this.

The touch-key interface can be powered-off when the screen is powered-off (and thus the power button non-functional) by writing "disabled" to sysfs.
`echo "disabled" > /sys/devices/platform/i2c-gpio.12/i2c-12/12-0020/power/wakeup`