tki-melfas: Update Touch-Key Interface to Modern Android Navigation Keys
Modern Android phones no longer have a Menu key. It has been replaced by a App-Switch key, AKA Recent Key. Userspace has support for this key, all we need to do is add the appropriate keycode to the kernel and map the key correctly.
To properly enable the Menu key, /system/usr/keylayout/tki_input.kl needs to be edited with a new entry like "key 248  APP_SWITCH"

In addition, Modern Android UIs have the Back key on the left, the Home key in the center, and the App-Switch key on the right.

Swap the order of the keys to conform to this standard. Obviously the physical labels on the phone are now not indicative of what the buttons do.