Only Call regulator_get() if the Regulator has not been Previous Acquired.
It is possible for a sound device to be opened multiple times for different functionalities, causing multiple calls to the function that enables the voltage regulators.

When the device is closed, the voltage regulator will be put and the pointer set to NULL. This means other outstanding pointers from other regulator_get calls caused by other openings of the device are now lost and cannot be have regulator_put() or regulator_disable() called on them.

To solve this issue, only obtain a regulator once with regulator_get() and call regulator_enable() and regulator_disable() multiple times on the same regulator object.