Revert "Switch default IO Scheduler to ROW"
This reverts commit 1bc0ee07532ad566b722ede622145c5209a043cb.

When connecting the phone to a network for the first time in a while, many syncing processes are running. The ROW Scheduler deals poorly with this and makes the UI nearly unusable. In fact, applications run so slowly that they are killed by the Android framework because they are non-responsive.

Note that this may be excaberated by the use of an external SD card and Link2SD to run most applications off of the external SD card.

The external SD Card is a Sandisk Ultra Class 10/UHS-I 32GB MicroSDHC card.