Make the display even darker
The previous attempt doesn't make the display dark enough for use in a blacked-out room. Here's another try.
It's likely that this is as dark as we can go without sacraficing too much image quality because gamma_sequence_n20[0].data[5] is 0 and cannot be lowered further.
Surprisingly, even at the dimmest setting, all steps of a 20-step grayscale ramp still can be seen.