Map Display Brightness Level 1 and 2 to the Lowest Level.
The CM11 Android Framework on the Pantech Burst allows a minimum brightness level of 10 (out of 255). In other words, when the brightness slider is set to the minimum, Android will write 10 to /sys/class/leds/lcd-backlight/brightness .

msm_fb_set_bl_brightness() maps this to the display brightness level (1-17). Unfortunately, 10 is mapped to 2, so we can't use the lowest brightness. To fix this, map both 1 and 2 to the lowest birghtness.