Revert "msm_tsens: Periodically raise interrupt when temperature is past trip threshold."
This reverts commit 449612fe660c9808a0fc02100fca46eaa3365c65.

Changed my mind - the original behavior of msm_tsens is good for implementing a
thermal throttling driver. My original thought was to use the TSENS_TRIP_STAGE2
interrupt to drive the thermal throttling driver instead of using a periodic
workqueue implementation, but it turns out that won't work because no interrupt
is generated when the temperature is below TSENS_TRIP_STAGE2 temperature. So,
the thermal throttling driver won't be kicked to reset the CPU speed back to
the maximum speed.

Instead we'll only have one interrupt when the temperature first crosses the
TSENS_TRIP_STAGE2 temperature, which starts up the periodic thermal throtting
workqueue. When the temperature falls below the TSENS_TRIP_STAGE1 temperature,
another interrupt will be generated which causes the msm_tsens driver to
re-enable the TSENS_TRIP_STAGE2 interrupt, so that we can be ready to kick-off
the the thermal throttling workqueue again when the temperature goes above
TSENS_TRIP_STAGE2.