msm cpufreq: Apply limit->allowed_{min,max} during msm_cpufreq_verify()
Before this commit, limit->allowed_{min,max} was applied during set_cpu_freq().
This doesn't work in all situations because set_cpu_freq() is only called when
the target cpu freq changes, and does not have any impact on the cpufreq policy.

If the CPU is operating outside of the new limits set by
msm_cpufreq_set_freq_limits(), and cpufreq_update_policy() is called to force
an update of the cpufreq policy, nothing will happen and the CPU will continue
at a out-of-limits frequency. This is because the cpufreq code thinks the
policy's new target frequency is the same as the current frequency.

To fix this, we need to force a change in the cpufreq policy. This is now done
in msm_cpufreq_verify().