msm_tsens: Use ordered versions of readl/writel instead of a memory barrier
Replace mb(), readl_relaxed(), and writel_relaxed() with readl() and writel().

In tsens_resume(), we want to make sure that the write to TSENS_THRESHOLD_ADDR
occurs before we enable the sensor with a write to TSENS_CNTL_ADDR. Otherwise,
the sensor may do a temperature reading and fire an unexpected interrupt because
the thresholds have not be initialized yet.

Using writel_relaxed() and mb() does not ensure this order.