msm_tsens_throttle: Throttle CPU when Screen is Off
When the screen turns off, the CPU will be throttled to prevent the phone from
getting hot while the user isn't looking at it or when it is in a pocket.

This might also save battery power if higher speeds are less efficent than lower
ones.

The default level of throttling is 1/2 of the highest speed. For our overclocked
Pantech Burst, this is 1026 MHz . The throttling level can be changed with the
screen_off_throttle_level kernel parameter.

Throttling requested via the cooling device interface (like via sysfs or by
thermal zones, like msm_tsens) will override the screen-off throttling when the
cooling device interface requests more throttling.