USB PHY Power Collapse: Add note about system hang problem
806c960c813 forced USB PHY Power Collapse for Presto. At the time no
negative consequences were observed, but all testing was done by
plugging in the host (computer or charger) end of the USB cable. The
device end of the USB cable was left connected at all times. In this
case everything works fine.

It turns out that plugging in the device end of the cable when the
device is suspended leads to hangs, so make note of this.

For now we'll keep USB PHY Power Collapse due to the power savings it
provides.