Refactor highmem-disable commit to allow build of highmem-enabled kernel
This refactors "1f526cec: Disable High Memory to Slightly Speed Up the
Kernel" so that it is possible to build a kernel with or without that commit's
changes.

This means the same kernel source can be used to build a highmem-enabled
kernel with a full 400MB of vmalloc memory, or a highmem-disabled kernel
with 128MB of vmalloc memory.

We were previously keeping a separate branch for this.